fix: adapt codebase to ruff 0.16 default rules
Sort imports, prefer collections.abc.Callable, drop redundant int(), and
simplify dotenv fixture string construction so CI passes on the bump.

+++ src/utils.py
@@ -85,7 +85,7 @@ def format_distance_km(distance_m: float) -> str:
 
 
 def format_duration_minutes(duration_s: float) -> str:
-    total_minutes = int(round(duration_s / 60.0))
+    total_minutes = round(duration_s / 60.0)
     hours, minutes = divmod(total_minutes, 60)
     if hours:
         return f"{hours}h {minutes}m"
